refactor(lotto-util): 로또 관련 유틸 기능 클래스 구현
