+++ src/main/java/lotto/LottoGenerator.java
@@ -4,13 +4,15 @@ import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
 import lotto.base_numbers.MainNumbersContainer;
+import lotto.lotto.Lotto;
+import lotto.lotto.LottoUtil;
 
 public class LottoGenerator {
 
     private static final int LIMIT_QUANTITY = 1000; // TODO 현재 메모리양 고려한 로직 구현
 
     public static List<Lotto> generateLottos(int paymentPrice) {
-        int count = Lotto.calculateLottoCount(paymentPrice);
+        int count = LottoUtil.calculateLottoCount(paymentPrice);
         validate(count);
         List<Lotto> result = new ArrayList<>();
         while (count > 0) {
